not sure correct usage when cmake is not present
headers are in opencv4/opencv2/...
I would expect either 4 or 2. the opencv docs suggest that
one should link using the opencv2/... pattern. This is likely
fine for cmake builds where the extra parent directory can be
encoded in the cmake config module. A little confusing for
users manually adding the includes directory for the compiler
frontend though.